Adicionado função para calcular notas recomendadas

+++ Prazo.cs
@@ -1,18 +1,18 @@
 public class Prazo{
     public Data diaDoPrazo;
     public string tipoDoPrazo;
-    private int notaTotal;
-    private int notaObtida;
-    private int notaRecomendada;
+    private float notaTotal;
+    private float notaObtida;
+    private float notaRecomendada;
     
-    public Prazo(Data diaDoPrazo,string tipoDoPrazo,int notaTotal,int notaObtida){
+    public Prazo(Data diaDoPrazo,string tipoDoPrazo,float notaTotal,float notaObtida){
         this.diaDoPrazo = diaDoPrazo;
         this.tipoDoPrazo = tipoDoPrazo;
         this.notaTotal = notaTotal;
         this.notaObtida = notaObtida;
     }
 
-    public int AdicionarNota(int nota){
+    public int AdicionarNota(float nota){
         if(nota>=notaTotal){
             Console.WriteLine("A nota obtida é maior doque a nota total");
             return -1;
@@ -22,7 +22,7 @@ public class Prazo{
         return 0;
     }
 
-    public int NotaObtida{
+    public float NotaObtida{
         get=>notaObtida;
         set{
             if(value<=notaTotal){
@@ -31,7 +31,7 @@ public class Prazo{
         }
     }
 
-    public int NotaRecomendada{
+    public float NotaRecomendada{
         get=>notaRecomendada;
         set{
             if(value<=notaRecomendada){
@@ -40,7 +40,7 @@ public class Prazo{
         }
     }
 
-    public int NotaTotal{
+    public float NotaTotal{
         get => notaTotal;
         set => notaTotal = value;
     }
